requirements.txt
Django==2.2.4
django-rest-framework==0.1.0
djangorestframework==3.10.2
pytz==2019.2
sqlparse==0.3.0
rsutils-django==0.1.3
Django==2.2.4
django-rest-framework==0.1.0
django-safedelete==0.5.1
djangorestframework==3.10.2
psycopg2-binary==2.8.3
pytz==2019.2
rsutils-django==0.1.3
sqlparse==0.3.0
Django==2.2.4
django-filter==2.2.0
django-rest-framework==0.1.0
django-safedelete==0.5.1
djangorestframework==3.10.2
psycopg2-binary==2.8.3
pytz==2019.2
rsutils-django==0.1.3
sqlparse==0.3.0
